move translator to new file.
