+++ mahjong/record/universe/translator.py
@@ -1,0 +1,51 @@
+from typing import TypeVar, Callable, Iterable, List
+
+from loguru import logger
+
+from mahjong.record.universe.command import GameCommand
+
+EventT = TypeVar('EventT')
+EventTransform = Callable[[EventT], Iterable[GameCommand]]
+
+
+class CommandTranslator:
+    def __init__(self):
+        self.defaults = []
+        self.defaults: List[EventTransform]
+
+    @staticmethod
+    def fallback_call(event, matchers: List[EventTransform]) -> List[GameCommand]:
+        for matcher in matchers:
+            value = list(matcher(event))
+            if value:
+                return value
+
+    # TODO: add preprocess and postprocess for event and command list. (to support timestamp attaching in tenhou).
+
+    def default_event(self, func: EventTransform) -> EventTransform:
+        self.defaults.append(func)
+        return func
+
+    def preprocess(self, event: EventT) -> EventT:
+        return event
+
+    def postprocess(self, event: EventT, command: List[GameCommand]) -> List[GameCommand]:
+        return command
+
+    def translate(self, event: EventT) -> List[GameCommand]:
+        pass
+
+    def __call__(self, event: EventT, strict=False) -> List[GameCommand]:
+        event = self.preprocess(event)
+        return_value = self.translate(event)
+        if return_value:
+            return self.postprocess(event, return_value)
+        return_value = CommandTranslator.fallback_call(event, self.defaults)
+        if return_value:
+            return self.postprocess(event, return_value)
+        else:
+            if strict:
+                raise ValueError("event <{}> is not transformed to game commands.".format(event))
+            else:
+                logger.warning("event <{}> is not transformed to game commands.", event)
+            return []
